GCFS: Fix out_of_bound exception during startup

+++ gcfs/src/gcfs_config_values.cpp
@@ -11,9 +11,9 @@
 
 std::string GCFS_ConfigValue::trimStr(const std::string& Src, const std::string& c)
 {
-	uint p2 = Src.find_last_not_of(c);
+	size_t p2 = Src.find_last_not_of(c);
 	if (p2 == std::string::npos) return std::string();
-	uint p1 = Src.find_first_not_of(c);
+	size_t p1 = Src.find_first_not_of(c);
 	if (p1 == std::string::npos) p1 = 0;
 	return Src.substr(p1, (p2-p1)+1);
 }
